Refactor web progress screen sections
